Refactor of the Type package.
The previous Type class API was far too complex. This refactor resulted
in a structure of the classes that is much simpler with utility methods
being extracted into a TypeUtil class.

The new structure results in a more generic implementation of the
various data structures (Trees, Lists, Graphs and Sets) that may be
used within Entity objects or directly in other objects.

The size of this refactor was larger than originally planned for, but
the results seem to be in accordance with what was planned. It should be
noted now that the Type interface is nothing more than a marker
interface now, with the number of methods it provides being reduced
drastically.

Signed-off-by: Gary Pampara <gpampara@gmail.com>