Corrected an index issue.
The mutation process within the MutationPositionUpdateStrategy uses a
process to randomly select a dimension from the current particle
position and then performs some modifications to mutation the value.

The error seems to be that the i-th dimension was being replaced and not
the randomly selected dimension (from which the original value is
determined).