Correct implementation of SortedList.
The previous implementation of SortedList was an implementation that
extended the java.util.LinkedList<E> class. This seemed very clunky and
as a result it was refactored to now be a composition class. Added some
more tests to verify the changes.

Removed the DynamicIterator. It is no longer used.

Minor updates to other classes.

Signed-off-by: Gary Pampara <gpampara@gmail.com>