Generalised FitnessCalculator.
The need to measure the fitness of a StructuredType<T> was needed for
the KMeans algorithm. It would not be a clean solution to create an
Entity adapter type object so that this functionality can be used.

It was rather decided to convert the current fitness calculators to be
generic and specialize the needed calculators based on the needed types.