Fix for when algorithms are set on StoppingConditions.
The Algorithm.addStoppingCondition() method internally made a call to
stoppingCondition.setAlgorithm(this).

While there is nothing wrong with the logic there, it breaks the with
some StoppingConditions because the object graph has not fully been
finalised when that Algorithm.addStoppingCondition() is invoked.

The code to apply the current algorithm to the StoppingCondition was
moved into the Algorithm.initialise() method to ensure that the object
graph is correct before any initialization or setup needs to occur.

Signed-off-by: Gary Pampara <gpampara@gmail.com>