Removed Entity properties in .equals and .hashCode
An interesting artifact was discovered when using a HashMap whereby the
same Entity was recorded multiple times in the map due to the properties
changing and as a result the hashCode() of the Entity changing.

The problem was due to the Entity classes requiring the properties of
the Entity to be checked for the hasCode() and equals() calculations.
This does not make sense because the same Entity may be maintaining
different candidate solutions during the process of an evolution.

Therefore, the correction was simply to remove the check on the metadata
properties of the Entity and only consider the identity of the Entity.

Signed-off-by: Gary Pampara <gpampara@gmail.com>