Correction to topology visitor classes.
Altered a few constructs in the TopologyVisitor classes. Removed the
ClosestParticleVisitor as the implementation was too specific to
Particles. This will be replaced by a more generic version of the class
that will work on all Entity objects and not just Particles.

Moved the notion of a "result" from the visitor to the subclasses.