Implementation of Graph equality.
Graphs are complex structures and as a result, testing the graph
structure equality (or isomorphism) can be very difficult and complex.
The important aspects for the comparison are the graph verticies and
edges. The verticies need to be connected in the same manner for both
graphs, although the layout and structure of the graphs may be different.