Removed Singleton pattern from DomainParser.
There was actually no need for the DomainParser to be managed with a
Singleton pattern. The object is simple enough to simply be instantiated
when needed and then thrown away.

This was changed to prevent potential issues as the parser is kept in
static address space once instantiated. This change may be pedantic, but
being catious never hurt before :P

Signed-off-by: Gary Pampara <gpampara@gmail.com>