Removed static RNG from util class.
The fact that a static member was present in the util class was an issue
that was overlooked. Utility classes should not be maintaining *ANY*
state at all. This could result in strange behaviour that might be very
difficult to trace.

Updated the javadoc of the classes to reflect the intention of the code
and to highlight any performance issues that may arise from using the
associated method too often. This is still naturally the user's
resposibility.

Signed-off-by: Gary Pampara <gpampara@gmail.com>