Modified the KMeans algorithm to add diversity.
A CentroidsDiversificationStrategy allows the KMeans algorithm to add
diversity to the centroids once it has been detected that it stagnated.
The NullCentroidsDiversificationStrategy is the default and does not add
any diversity, i.e. adhering to the standard KMeans algorithm.

The StandardCentroidsDiversificationStrategy takes the 'interval'
and 'diversifyRatio' as parameters. In other words, diversity is added to
the specific centroid only when 'interval' iterations have passed
without any change occuring for that centroid. The diversity that is
added is achieved by the VectorUtils.jitter() method.

Extra helper methods were added in ClusteringUtils to assemble and
disassemble centroids.

Extra methods were added in the Vectors utility class to add noise
(jitter) to a vector as well as set new bounds for a vector.

A scale() method was implemented in the Maths utility class, but as I
continued development I realised that it isn't really necessary for
what I am doing. Nonetheless, it is still a valid addition. A unit test
for the scale() method failed, because the Maths.EPSILON value was too
small. It was therefore changed from 1.0E-15d to 2.0E-15d.

I also added the PlotClustersAndCentroids measurement that is useful
when debugging 2D clusterings using GNUplot. It can also serve as an
example for other people to see how to use GNUplot as debugging tool in
CIlib.

JavaDoc comments were added and modified to explain each addition and
modification respectively.