Games: Added GameSeedingStrategy and subclasses.
A seed is now only set once when a game is played at the beginning of
the game. This makes it possible to reproduce specific games. The class
contains a ranomzer for the games framework and the ability to set seed(s)
using specified strategies. It is possible to only play a fixed game during
evaluation, or a fixed set of games. It is also possible to generate a number
of seeds to use for a number of iterations, and to re-generate that list at
fixed intervals.

Signed-off-by: Leo Langenhoven <llangenhoven@cs.up.ac.za>
Signed-off-by: Gary Pampara <gpampara@gmail.com>