Games: State based games framework.
Completed the framework that optimizes games by generating the game
tree and then selecting the best leaf node with a heuristic funcion.

This consists of the stateEvaluationAgent class,
the StateTraversalStrategy class and the StateEvaluation
class.

The StateTraversalStrategy class dictates how the game's state space is
explored and it selects the best state by using the stateEvaluator class
as its node evaluator. An implimentation of Min-Max with alpha-beta
pruning is supplied as well as an EndGame State Evaluator (that simply
evaluations a completed game with win lose draw scores) and a
NeuralStateEvaluator that uses a NeuralAgent to evaluate a game state.

Signed-off-by: Leo Langenhoven <llangenhoven@cs.up.ac.za>
Signed-off-by: Gary Pampara <gpampara@gmail.com>