Games: Added Agent measurement framework.
The agent measurement framework is used to record any behavior by an
agent during the course of the game. At the end of each game iteration
each agentmeasure class supplied to the game can record information
about the current state of the game. This is used so that
gamescoringstratgies that need specific information (for example how
many times an agent performed a certain task) has access to that information.

Signed-off-by: Leo Langenhoven <llangenhoven@cs.up.ac.za>
Signed-off-by: Gary Pampara <gpampara@gmail.com>