Removed getSocialBestFitness() from Entity.
The social best fitness construct is very specific to PSO and as a
result does not really make sense for all entities.

This change can be reverted - but for now I think that it is valid.

All the social best fitness ideas where pushed down into classes related
to Particle.

Signed-off-by: Gary Pampara <gpampara@gmail.com>