Unit test for PSO algorithm.
By setting the seed of the PRNGs, we are able to test stochastic
algorithms. The initialisation of the algorithm will result in a state
that can be predetermined.

Running the algorithm in this fashion will always produce the same
output. This results in a deterministic manner to test the stochastic
algorithm.

Signed-off-by: Gary Pampara <gpampara@gmail.com>