Corrected boundary constraints to be generic.
The current boundary constraints were very specific to either Particles
or simply oblivious to the type of entity to which they operated on.

The refactor made the constraints a little more generic, however, there
are some constraints that are dependent on the fact that a velocity is
available within the entity. These constraints will throw an exception
if the required data is not available.

Signed-off-by: Gary Pampara <gpampara@gmail.com>