Entity selection based on Comparator.
There has been the requirement that the manner in which the required
Entity is selected from a Topology needs to be dynamic. This could
create large amounts of confusion, but thankfully, passing in a
Comparator instance should resolve this issue. It is now the user's
responsibility to provide the required type of Comparator when the
selection is to be _different_ from the normal mode of operation.

Signed-off-by: Gary Pampara <gpampara@gmail.com>