Corrected Entity Comparators.
The Entity comparators, especially the AscendingFitness and
DescendingFitness comparators were not very clear. The changes with the
unit tests make the usage of the tests more clear.

Cleanup of a few classes where made to make the intention of the code
much clearer as well as removing some redundant code.

Signed-off-by: Gary Pampara <gpampara@gmail.com>