Changes required to existing classes.
Some of the classes got removed to accomidate the new Multi-objective
structure:

This meant that the Guide selection strategies got removed and replaced
by guide selection strategies in the pso package. (Guide selection is a
pso concept and not a general Multi-objective concept).

MOPSO got removed: mutli-objective PSO's will get constructed by adding
different iteration strategies and configuring different guide selection
strategies (construction of MOPSOs will be modular).

The classes that got modified are:

EntityType got modified with additional fields corresponding to Global
and Local guides for PSO.

The Archive interface got proper methods to reflect the storage of multi-
objective solutions -- the constrained and unconstrained archives will
be added in the next commit.

The VEPSO Velocity Update class will be replaced by a general MO Velocity
Update Strategy that applies to both VEPSO as well as some other MOPSOs.

The ExpressionFunction class also changed to work a little more
intuitive (from a logical perspective). Thus, your expressions that you
can set are a little more flexible than the way it worked previously.

Some changes were made to the MOOptimisationProblem class to enable it
to behave like a list of Optimisation Problem classes (it implements
java's list iterface).

The MOFitness class got additional methods like equals etc. to enable it
to work with the Arhive classes (especially the SetBased implementation
that will be commited). This also caused some changes to the
MOFitnessTest class.

OptimisationSolution got equals() and hashCode() methods to enable
instances of this class to be stored in the Set-based implementation of
an archive.

Signed-off-by: Gary Pampara <gpampara@gmail.com>