New MOPSO classes added.
These classes implement a generic framework for most Multi-objective
PSOs. An implementation for VEPSO is provided with an example xml file
(vepso.xml) to illustrate how VEPSO should be set up. Some generic
classes are also included to handle the storage of multi-objective
solutions in an archive (these classes should be applicable not only to
MOPSOs but Multi-objective optimisation in general). Classes were also
added to measure the Pareto optimal set as well as the Pareto optimal
front at each iteration.

Things that are missing: The different measurements to determine how
well a multi-objective optimisation algorithm is performing. Need to
discuss how we are going to implement this.

See the javadoc's of the various classes for a better explanation of
how these classes are used.

Signed-off-by: Gary Pampara <gpampara@gmail.com>