Some VEPSO and MOPSO bugfixes and cleanup.
Removed the CompositeIterationStrategy class in favor of
IterationStrategy classes that wraps other IterationStrategy classes.
The class that follows this new convention is the
ArchivingIterationStrategy class that wraps another IterationStrategy
class (like SynchronousIterationStrategy etc.). Renamed
ArchivingIterationStep to ArchivingIterationStrategy.

Changed the Archive class to only make use of the add and addAll methods
to insert solutions into the archive (removed the accept methods). This
change spilled to the ConstrainedArchive and SetBasedConstrainedArchive
classes as well as the QuadTree placeholder class. Other classes that
also changed is the ArchiveTest and ConstrainedArchiveTest classes.

Fixed a bug in the ArchiveReevaluationResponseStrategy class by
uncommenting the step that adds all solutions to the archive. (Required
for when the archive changed as a result of the environment change).

Removed the GuideSelectionIterationStep and moved the GuideSelectionStrategy
and GuideUpdateStrategy classes out of this class into the
MOVelocityUpdateStrategy class (a bit cleaner code and less classes).

The abovementioned changes resulted in changes to the vepso.xml and
dynamic-vepso.xml files.

Signed-off-by: Gary Pampara <gpampara@gmail.com>