Corrected ordering of Entity objects.
The ordering of Entity instances was being double negated in some
instances. The comparison logic for Entity instances is performed within
the Fitness instance associated with the Entity.

As a result of this, it is not necessary to alter the direction of the
sorting and a default sorting which is descending (when looking at the
fitness objects) is viable.

This also is needed to correct a current issue within the GA
implementation.

Signed-off-by: Gary Pampara <gpampara@gmail.com>