Niching algorithm fixes and unit tests.
Unit tests have been added to ensure the correctness of the various
components of a niching algorithm. Currently the niching algorithm
implementation is the NichePSO, however, injecting the correct members
into the Niche will result in a different algorithm.

Updates to many javadocs. This should aid in the API.

Several bug fixes in the Niching code have been corrected and are
verified by the test cases.

References for the NichePSO have been added in the javadoc.

Signed-off-by: Gary Pampara <gpampara@gmail.com>