[Games] Updated Game Seeding Strategy
Added the functionality to the game seeding strategy code so that one
can do the following:

Specify a list of seeds, an iteration modulus and a number of seeds to
use. This enables you to use a number of seeds from a list, say 2, at a
time for a fixed period. Afterwords you switch to the next 2 seeds in
the list and so forth.

Signed-off-by: Leo Lengenhoven <leolangen@gmail.com>
Signed-off-by: Gary Pampara <gpampara@gmail.com>