Rename methods to make intention clear.
The ChangeStreategy classes were defined with a change(Problem) method.
This kind of naming doesn't read well and the intention of the method is
lost. Renamed to reestablish the intention of the method.

Updated the javadoc for the problem.changestrategy package. Added
package level documentation as well.

Bumped the version to 0.7.1-SNAPSHOT

Signed-off-by: Gary Pampara <gpampara@gmail.com>