Removed dependency on Random within MathUtil.
The management of static state within MathUtil is not very attractive
and as a result, causes many issues with unit tests.

The state has been removed and modifications to remaining classes have
been made. Added the Randomizable interface to add randomizable support to
StructuredTypes and Numerics. The randomization process now requires a
provided random number generator to be provided.

Signed-off-by: Gary Pampara <gpampara@gmail.com>