A generic and fluent selection interface.
Selection is a process that is very important within CI, with special
use in ECs.

This patch is a combination of work to reflect such an implementation.
As a result of this new interface, all selection operations will now be
generic. The interface has the ability to accept any form of list for
input and correctly chain a variety of operations together to achieve
the desired selection operation.

Various small corrections have been made to other classes as a result of
these patches.

A new set of classes have been introduced to allow for defined
selection "recipes" to be applied. These recipes include:
 - Tournament selection
 - Roulette wheel selection
 - Rank-based selection
 - Random selection
 - etc.

Signed-off-by: Gary Pampara <gpampara@gmail.com>