Vector for Numeric types.
Changed the manner in which the type Vector operates. Vectors are generally
associated with mathematical operations and as a result, the Vector has
now been made a specialization of AbstractList<E> with Numeric as the
type parameter.

For general object containment in a list type structure, please use the
TypeList class which has been added.

Corrected a bug where the upper bound was being excluded in Vector and
TypeList.

Signed-off-by: Gary Pampara <gpampara@gmail.com>