Removal of ambiguous getPopulationSize().
Obtaining the size of the population was possible in two ways. However,
using the getPopulationSize() method was flawed as it returned the size
of the initial population and not the current population.

This method has been removed and you should be using topology.size()
instead. To get the size of the initial population, the
PopulationInitialisationStrategy should be queried for the entity
number.

Signed-off-by: Gary Pampara <gpampara@gmail.com>