Corrected Constriction PSO.
Corrected the ConstrictionVelocityUpdate: The constriction coefficient has
to be calculated only once during the course of the algorithm (as opposed
to every time it is used) similar to inertia.

The calculation also uses c1 and c2 WITHOUT the randomization of r1 and r2.
Resulting implementation delivers good results that are very similar to
those reported by reference papers.

Signed-off-by: Gary Pampara <gpampara@gmail.com>