Generalization of function types.
Functions provide a relationship for an input to an output. These
changes define that the input and output types are parameterized.

This removed the need to have:

    public Object getMinimum() {
        return new Double(0.0);
    }

and simply allows:

    public Double getMinimum() {
        return 0.0;
    }

where the function is defined to be a Function<F, Double> where F is any
input type.

Additionally, this makes the API simpler to use and adds more type
safety to the functions. The most notable is that DiscreteFunctions now
work with Integer objects and not Double objects, was was the case
previously.

Signed-off-by: Gary Pampara <gpampara@gmail.com>