Bee position update unit test fix.
Fixed bug that caused the unit tests of VisualPositionUpdateStategy
and ExplorerBee to fail, the issue was related to bees leaving the
function domain, which should be allowed. The tests were also expanded.

Removed dependency of ExplorerBee on Algorithm.get(), enabling simpler
testing.

Changed exactly when bees calculate fitnesses, new implementation is
slightly more efficient. Changed a number of method names and made slight
changes to the abc.xml file.

Signed-off-by: Gary Pampara <gpampara@gmail.com>