Prevent thread local variable polution.
These changes prevent the leaking of ThreadLocal variables. Previously
the simulation object graphs were created ahead of time and then
executed in threads. This effectively resulted in the ThreadLocal
variable being placed on the main thread and _not_ the intended
simulation thread. Yes, this is BAD.

Created a logical separation and ensured that the object graph
construction occurs within the simulation thread and not in the main
thread.

Tested-by: Wiehann Matthysen <wcmatthysen@gmail.com>
Signed-off-by: Gary Pampara <gpampara@gmail.com>