Add type information for measurements.
Measurements simply used to return a 'Type' instance - this was rather
restrictive. The type of the return should be tailored to the manner in
which the measurement is used.

Measurement has been altered to be Measurement<E extends Type> in order
to correct this. Having a better more type safe implementation is of course
better than a non-typesafe implementation.

Signed-off-by: Gary Pampara <gpampara@gmail.com>