Minor DE performance updates.
Corrected the replacement of the parent entity within the DE. It is
faster to replace the current pointer with a new entity that is already
built and available, than to replace the contents within the parent with
that of the offspring.

Additionally updated the naming in the GBestTopology to be more generic.

Signed-off-by: Gary Pampara <gpampara@gmail.com>