Added Heterogeneous Cooperative Algorithm
Heterogeneous cooperative algorithms use a combination of different
PBA's in the optimization process. This implementation re-distributes
the problem vector amongst the participating algorithms at certain
intervals. The re-distribution is performed by a redistribution
strategy, which allows for a variety of approaches. Tests for these
classes are included.

Some of the new cooperative framework class's comments are cleaned up,
and this patch also includes a test for the
CooperativeCoevolutionAlgorithm class.

Signed-off-by: Gary Pampara <gpampara@gmail.com>