Removed redundancy in SocialEntity
SocialEntity is another interface that an Entity may implement to cater
for the social interaction of a collection of Entity objects.

SocialEntity was extending Entity, which is wrong. A social entity is a
normal entity that has social properties, such as in the case of the
Particle or even an Ant. These specific Entity instances can implment
the interface to enable the social behavior.

Signed-off-by: Gary Pampara <gpampara@gmail.com>