Fix Cooperative Algorithm
Fixed cooperative algorithm to test for ZeroContributionSelectionStrategy
strategy instead of null. If a participating algorithm's contribution strategy
has not been set (if it is an instance of ZeroContributionSelectionStrategy
strategy) when the algorithm is added to the coevoltuion algorithm, then
it is given the same instance that the coevolution algorithm contains.
Otherwise the strategy is unchanged.

Signed-off-by: Gary Pampara <gpampara@gmail.com>