Remove redundant UniqueSelection.
The UniqueSelection class created a lot of redundant code. In order to
address this, the logic related to the random selection in the normal
and unique cases was pulled out into a set of classes and interfaces. As
a result, the desired internal operation was altered to ensure that the
functionality was correct, transparent to the outside world.

Signed-off-by: Gary Pampara <gpampara@gmail.com>