Simplification of stopping conditions
An interesting idea was presented by Wiehann regarding the manner in
which the stopping conditions for algorithms have been implemented. The
idea is to allow the stopping conditions to be Predicate<T> instances.

This alteration should hopefully allow for future changes regarding
the application of combinatoric logic to stopping conditions.

The fact that the code becomes simpler is another big bonus :P

Signed-off-by: Gary Pampara <gpampara@gmail.com>