Selection-EDSL cleanup
The EDSL used within the Selection API is rather ambiguous. For example,
it was possible to call unique() multiple times on the same Selection
object.

In order to address this, all Syntax interfaces have been either dropped
or renamed to SelectionBuilders with a clear definition of how method
traversal for the Selection-EDSL.

The selection methods (first, first(n), last and last(n)) have been
removed in favor of a SamplePredicate that may be applied to the current
selection sample to isolate the required selection subset. This
predicate can be used via the select() method in the EDSL. Additionally,
this does open up the possibility of other predicates as only first() and
last() have been provided.

Signed-off-by: Gary Pampara <gpampara@gmail.com>