JUnit 4.7 in central maven repo.
The updating of the central maven repo with the junit artifacts seems to
be lagging. As a result the version of junit needs to be bummped from
4.8.1 to 4.7.

There is an issue for Kent Beck on his Github site:
    http://github.com/KentBeck/junit/issues/issue/66

It seems that the next version, whereby the update to the central maven
repo will happen, is version 4.9.

Signed-off-by: Gary Pampara <gpampara@gmail.com>