Assign unique MeasurementSuite per Simulation.
The main problem with the manner in whcih the Simulator processes data
is that the data causes the threads to block and wait for slower threads
to catch up. This has several disadvantages which are not required.

This patch alters the manner in which data is persisted, by allowing
each individual simulation thread to stream it's observed measurements
to a file on disk directly. This manner of output alleviates any form of
thread contention and synchronization.

At the end of the run, the temporary files would need to be read and
merged into a final results file. The merging functionality is not yet
available, but will be included in a subsequent patch.

Signed-off-by: Gary Pampara <gpampara@gmail.com>