Removed state from ClusteringUtils.
All the changes in this commit is related to removal of the objects in
ClusteringUtils that caused it to have a state. ThreadLocal is now
removed and the important methods are static.

Another important addition is the Cluster<C extends Vector> class is
also a Type, more specifically a Set<Pattern<C>>.
The Pattern<S extends Vector> class is also now a Type, more
specifically a ForwardingStructuredType<S>.

All the other changes are related to incorporating the above two points.