Lower bound should be -Double.MAX_VALUE.
The lower bound of the Linear activaton function was set to
Double.MIN_VALUE, which according to the JDK is "a constant holding
the smallest positive nonzero value of type double". Since the Linear
activation function is unbounded, the lower bound should actually be
negative Double.MAX_VALUE, i.e. -Double.MAX_VALUE.

Signed-off-by: Theuns Cloete <theuns.cloete@gmail.com>
Signed-off-by: Gary Pampara <gpampara@gmail.com>