Output directive to output data with format.
Modifications to the manner in which data is collected required that the
final results final be moved out of the measurements details.

The resulting implications are that the 'simulation' tags now specify
the number of samples and that the simulation defines an 'output' tag
for the final output.

The final output can be performed in a number of different data formats.
namely: TXT, CSV or XML. Currently, only the TXT scheme is implemented.

Signed-off-by: Gary Pampara <gpampara@gmail.com>