Vector randomization through Builder.
The mutation of a Vector instance is not valid as the randomize()
operation is really related to the manner in which a Vector instance is
created.

The randomization has been moved to the Vector.Builder in the interests
of ensuring that the creation logic is centralized in a single place.

Signed-off-by: Gary Pampara <gpampara@gmail.com>