Deprecate getReal()
The method name "getReal()" doesn't really indicate what the result of
the method invocation will be. A clearer naming is required that will
ensure that the intention of the method is conveyed.

getReal() should be replaced with a new method called
"doubleValueOf(index)" in the case of the Vector and "doubleValue()" in
the case of Numeric instances.

Signed-off-by: Gary Pampara <gpampara@gmail.com>