Particle position and velocity noise decorators.
Decorators are supplied to add noise, sampled from a gaussian
distribution, to a particle position and velocity updates.

A new xml file, noisyGbestPSO.xml is also provided to show how the
new decorators can be used to add noise to particle updates.

The decorators should eventually be changed to sample noise from
any probability distribution, with a given mean and deviation.
Currently this is not possible because of the way the RandomNumber
class is implemented.

Signed-off-by: Gary Pampara <gpampara@gmail.com>

Update Noisy particle decorators to be simplified.

The implementation of the noisy decorators was a little complex. Using
some composition we can simplify it tremendiously.

Signed-off-by: Gary Pampara <gpampara@gmail.com>