Remove operator pipeline.
The construct of a pipeline resulted in some usage problems with the EC
algorithms. To correct this, the pipeline has been removed and the GA
has been prepared to be differentiated into 'generational' and
'steady-state'.

If needed this idea can be added again later, but for now, it's simply
getting in the way.

Signed-off-by: Gary Pampara <gpampara@gmail.com>