Remove Pair class.
The Pair class was a general container for two objects. This was rather
handy, but it did not provide a clear purpose of what it was being used
for. Seeing as that it was only used in very select areas, special
containers have been created privately to achieve the same result, with
better context of their intention.

Signed-off-by: Gary Pampara <gpampara@gmail.com>