Updated license information to be LGPL.
Updated the license header to include the required LGPL information.
CIlib has moved over to the LGPL in the hopes that people will consider
using the library more now, especially since the potential problem with
GPL forcing people to release their applications as GPL as well. This
change was spurred on by numerous requests from users.

Signed-off-by: Gary Pampara <gpampara@gmail.com>