Update of Dynamic particle re-evaluation.
This is a fix to the dynamic particle evaluations. The _entire_ dynamic
framework for PSO needs to be redesigned as currently it does not make
sense at all.

I'm applying this patch, but I'm not happy about the quality of this
framework.

Furthermore, this code _will_ see a redesign - and VERY soon.

Signed-off-by: Gary Pampara <gpampara@gmail.com>