Population diversity measure using entropy.
An entropy measurement is provided to calculate population entropy
in each dimension of the search space. Entropy is measured by first
dividing the search space into a number of intervals. The probability
that a given entity in the population will fall within a specific
interval is then calculated and use to obtain an entropy value for
each dimension.

The entropy values are used to measure the diversity of the population
in one of two ways. In the first case, the diversity is calculated as
the average entropy over all dimensions. In the second case, the
diversity is taken as the maximum entropy over all dimensions.

Signed-off-by: Gary Pampara <gpampara@gmail.com>