Refactored clustering problem and functions.
The ClusteringFitnessFunction idea was refactored so that the clustering
functions should make use of composition instead of inheritance. For
this, the ClusteringFunction interface was introduced and the
ClusteringFitnessFunction class was renamed to the
ClusteringErrorFunction. All the "helper methods", that were always
inherited, were refactored into their own classes. This class is now
only responsible for handling the getter and setter methods and the
evaluate() method should not be used anymore. The apply() method of the
ClusteringFunction interface has a dependency injection "feel" and
requires all the necessary information to calculate the clustering error.

Added the InverseErrorFunction. This name might not be appropriate, but
we can change it later.

The ClusteringProblem class was renamed to PartitionalClusteringProblem,
because the ClusteringProblem interface was added. The
PartitionalClusteringProblem was refactored and made responsible to pass
all the requirements to the configured ClusteringFunction's apply
method. Instead of wrapping a FunctionOptimisationProblem that can
either minimise or maximise the output of a function, the
PartitionalClusteringProblem now makes use of the new FitnessFactory
idea that can either be a MinimisationFitnessFactory or
MaximisationFitnessFactory that returns a Minimisation- or
MaximistionFitness respectively. Wiehann and I discussed a better way of
doing this, but the current solution will suffice until then. The zMax
is also now calculated correctly and calculated and stored in the
PartitionalClusteringProblem instead of being calculated somewhere else.

The ClusteringUtils class was renamed to the ClusteringFunctions helper
class and refactored to be a lot simpler.

All aggregated functions were moved to the
functions.clustering.aggregated package.

Unit tests have been added for all the apply() methods implemented in
each clustering error function. The unit tests were much simpler to
implement since the apply() method has a dependency injection "feel".