Additional probability distributions.
Three new probability distributions (Laplace, Exponential & Gamma) are
provided. The ProbabilityDistributionFunction interface was modified to
accomodate a variable number of control parameters for sampling random
numbers from a distribution.

Supplied a new benchmark function (Elliptic).

It will be determined with the usage of the API if the changes to the
ProbabilityDistributionFunction interface make sense. I'm concerned about
the misuse of the varargs here and would prefer a solution that is more
elegant. (Gary Pampara)

Signed-off-by: Gary Pampara <gpampara@gmail.com>