Selection framework simplification.
A lot of focus has been placed into making the selection process more
intuitive and simpler. As a result, this refactor reduces the complexity
of the selection classes and promotes a simpler API.

Albeit that this code is far from complete, it is a semi-decent point to
start form as more optimizations will be added, given time.

Signed-off-by: Gary Pampara <gpampara@gmail.com>