Framework for Heterogeneous PSO (HPSO).
The DynamicHeterogeneousIterationStrategy makes use of a
BehaviorChangeTriggerDetectionStrategy to detect when a particle should
change its behavior. When a change must occur, a new ParticleBehavior
is selected from a predefined pool of behaviors.

To simplify the behavior selection process, a ParticleBehavior object
encapsulates the velocity and position update strategies that a particle
uses.

The HeterogeneousPopulationInitialisationStrategy is used to populate
the topology with particles that follow different behaviors from the
start of the algorithm. Using this inititialisation strategy without
the DynamicHeterogeneousIterationStrategy results in a static HPSO.

The DynamicHeterogeneousIterationStrategy can be used in combination
with the DynamicIterationStrategy to facilitate HPSO in dynamic
environments.

Signed-off-by: Gary Pampara <gpampara@gmail.com>