Remove RandomizingControlParameter
The RandomizingControlParameter has been removed in favour of an
approach that uses RandomProvider(s) in the velocity update along
with other ControlParameters. The new approach is more transparent
and results in code closer to published work. Some general clean-up
of the velocity updates involved has also been done. Primarily the
BareBones velocity now makes better use of inheritance avoiding a
lot of reimplementation.

Signed-off-by: Gary Pampara <gpampara@gmail.com>