Ignore specified columns during pattern conversion.
The PatternConversionOperator was not very flexible in skipping over
unwanted columns of a data set. This commit adds a feature to ignore a
set of specified columns.

Added the PatternConversionOperator.ignoreColumnIndex(int) method that
can be used to specify the column indices that shoud be ignored. This
method can be called multiple times. It can also be used in XML as
follows:

<addDataOperator class="io.transform.PatternConversionOperator"
                 classIndex="10" classLength="1">
    <ignoreColumnIndex>0</ignoreColumnIndex>
    <ignoreColumnIndex>7</ignoreColumnIndex>
    <ignoreColumnIndex>3</ignoreColumnIndex>
</addDataOperator>

Refactored the PatternConversionOperator.operate(DataTable) method to be
cleaner.

Refactored the two existing unit tests. Added five more unit tests.