Fixed static ThreadLocal variables bug.
Firstly, because we are now using Java's threadpool-type approach to
managing our threads in a simulation we get a weird side-effect. The
thread executing the algortihm lives beyond the lifetime of the
algorithm and subsequently causes the threadlocal variables to remain.
This had the effect of the Archive of one simulation living on to be
shared in the next simulation that the thread would execute, causing the
Archive to grow and deceiving simulation results to be reported. As a
temporary fix (until we can remove the static state) the ThreadLocal
variables are explicitly destroyed at the end of an algorithm's
execution. Threalocal variables that got affected: AbstractAlgorith,
EntityIdFactory and Archive.

Secondly, made some cleanups to a couple of test cases where Vector
builders are used intead of the add method on the Vector class etc.