Core executor structure correction.
The manner in which the executor operated was not preferable. To correct
this issue, the way in which the executor obtained the algorithm,
initial population and the stopping conditions needed to change.

The initial population is no longer created for the user. The user is
now responsible for the initial population creation. This not only helps
in making the unit testing simpler, but it would give the user much more
flexibility than what would be the case if CIlib created the population.

Signed-off-by: Gary Pampara <gpampara@gmail.com>