Ring-based population selection fix.
The RingBasedPopulationSelector class selects the next algorithm from
the total selection of available algorithms (in a multi-population based
algorithm). The current selection code (Arrangement) assumed that the
list that it selects from is a list of comparable algorithms which
caused the code to break as Algorithm doesn't implement comparable. This
restriction was removed and only the Arrangement classes (such as
SortedArrangement) that requires the list to be comparable enforces this
restriction now.