Problem defined as a function and Fitness removed.
Defining a problem as a function enables some interesting combinations
and simplified the class hierarchy. Allowing for this is required for
more complex types of problems.

Additionally, the Fitness object has been removed in favor of a using an
Option<Double> in conjunction with a FitnessComparator to actually do
the comparisons. This change allows for a simplification and allows the
logic for comparisions to be injected where needed and is not on the
Fitness objects themselves.

Signed-off-by: Gary Pampara <gpampara@gmail.com>