Annotation based eventing.
Raising events should not be a complex thing to do. As a result, the
idea was suggested that annotations be used to annotate classes to
indicate that a specific method could possibly raise some events.

The eventing is optional, and is not something that is enabled per
default. The user may enable eventing by including a Guice module when
creating an Injector instance.

All methods that raise an event should have the following annotation
applied:

    @Raises(EventType.class)

The EventType is needed to determine which event is raised. The standard
configuration provided two event types that are enabled defaultly,
namely: IterationEvent and FitnessEvaluationEvent

The default events may be disabled when configuring the Guice module
builder, invoking the disableDefaults()

Signed-off-by: Gary Pampara <gpampara@gmail.com>